CY2008 added a new class _swiss.py, pep8 pass
