+++ doc/gsim/chiou_youngs_2008.rst
@@ -7,11 +7,3 @@ Chiou and Youngs 2008
 .. autoclass:: ChiouYoungs2008
     :members:
 
-.. autoclass:: ChiouYoungs2008SWISS01
-    :members:
-
-.. autoclass:: ChiouYoungs2008SWISS06
-    :members:
-
-.. autoclass:: ChiouYoungs2008SWISS04
-    :members:
