+++ doc/gsim/chiou_youngs_2008_swiss.rst
@@ -1,0 +1,14 @@
+==================================
+Chiou and Youngs 2008 Swiss Models
+==================================
+
+.. module:: openquake.hazardlib.gsim.chiou_youngs_2008_swiss
+
+.. autoclass:: ChiouYoungs2008SWISS01
+    :members:
+
+.. autoclass:: ChiouYoungs2008SWISS06
+    :members:
+
+.. autoclass:: ChiouYoungs2008SWISS04
+    :members:
