+++ openquake/hazardlib/gsim/chiou_youngs_2008_swiss.py
@@ -1,0 +1,109 @@
+# The Hazard Library
+# Copyright (C) 2012-2014 GEM Foundation
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+"""
+Module exports
+:class:`ChiouYoungs2008SWISS01`,
+:class:`ChiouYoungs2008SWISS06`,
+:class:`ChiouYoungs2008SWISS04`.
+"""
+from __future__ import division
+
+import numpy as np
+
+from openquake.hazardlib.gsim.base import GMPE, CoeffsTable
+from openquake.hazardlib import const
+from openquake.hazardlib.imt import PGA, PGV, SA
+from openquake.hazardlib.gsim.chiou_youngs_2008_swiss_coeffs import (
+    COEFFS_FS_ROCK_SWISS01,
+    COEFFS_FS_ROCK_SWISS06,
+    COEFFS_FS_ROCK_SWISS04)
+from openquake.hazardlib.gsim.chiou_youngs_2008 import (ChiouYoungs2008)
+from openquake.hazardlib.gsim.utils_swiss_gmpe import _apply_adjustments
+
+
+class ChiouYoungs2008SWISS01(ChiouYoungs2008):
+
+    """
+    This class extends :class:ChiouYoungs2008,
+    adjusted to be used for the Swiss Hazard Model [2014].
+
+    1) kappa value
+       K-adjustments corresponding to model 01 - as prepared by Ben Edwards
+       K-value for PGA were not provided but infered from SA[0.01s]
+       the model considers a fixed value of vs30=1100m/s
+
+    2) small-magnitude correction
+
+    3) single station sigma - inter-event magnitude/distance adjustment
+
+    Disclaimer: these equations are modified to be used for the
+    Swiss Seismic Hazard Model [2014].
+    The use of these models in other models
+    is the soly responsability of the hazard modeler.
+
+    Model implmented by laurentiu.danciu@gmail.com
+    """
+
+    def get_mean_and_stddevs(self, sites, rup, dists, imt, stddev_types):
+
+        mean, stddevs = super(ChiouYoungs2008SWISS01, self).\
+            get_mean_and_stddevs(sites, rup, dists, imt, stddev_types)
+
+        log_phi_ss = 1
+        tau = super(ChiouYoungs2008SWISS01, self).\
+            get_tau(ChiouYoungs2008.COEFFS[imt], rup)
+
+        ln_y_ref = super(ChiouYoungs2008SWISS01, self).\
+            _get_ln_y_ref(rup, dists, ChiouYoungs2008.COEFFS[imt])
+
+        exp1 = np.exp(ChiouYoungs2008.COEFFS[imt]['phi3'] *
+                      (sites.vs30.clip(-np.inf, 1130) - 360))
+
+        exp2 = np.exp(ChiouYoungs2008.COEFFS[imt]['phi3'] * (1130 - 360))
+
+        nl = super(ChiouYoungs2008SWISS01, self).\
+            get_nl(ChiouYoungs2008.COEFFS[imt], ln_y_ref, exp1, exp2)
+
+        mean, stddevs = _apply_adjustments(
+            ChiouYoungs2008.COEFFS, self.COEFFS_FS_ROCK[imt], 1,
+            mean, stddevs, sites, rup, dists.rjb, imt, stddev_types,
+            log_phi_ss, NL=nl, tau_value=tau)
+
+        return mean, stddevs
+
+    COEFFS_FS_ROCK = COEFFS_FS_ROCK_SWISS01
+
+
+class ChiouYoungs2008SWISS06(ChiouYoungs2008SWISS01):
+
+    """
+    This class extends :class:ChiouYoungs2008,following same strategy
+    as for :class:ChiouYoungs2008SWISS01 to be used for the
+    Swiss Hazard Model [2014].
+    """
+
+    COEFFS_FS_ROCK = COEFFS_FS_ROCK_SWISS06
+
+
+class ChiouYoungs2008SWISS04(ChiouYoungs2008SWISS01):
+
+    """
+    This class extends :class:ChiouYoungs2008,following same strategy
+    as for :class:ChiouYoungs2008SWISS01 to be used for the
+    Swiss Hazard Model [2014].
+    """
+
+    COEFFS_FS_ROCK = COEFFS_FS_ROCK_SWISS04
