+++ openquake/hazardlib/gsim/utils_swiss_gmpe.py
@@ -30,7 +30,6 @@ def _compute_C1_term(C, dists):
     idx = dists < C['Rc11']
     c1_dists[idx] = C['phi_11']
     idx = (dists >= C['Rc11']) & (dists <= C['Rc21'])
-    idx = (dists >= C['Rc11']) & (dists <= C['Rc21'])
     c1_dists[idx] = C['phi_11'] + (C['phi_21'] - C['phi_11']) * \
         ((dists[idx] - C['Rc11']) / (C['Rc21'] - C['Rc11']))
     idx = dists > C['Rc21']
